Error handling and memory allocation

+++ em.maksim/P4/main.cpp
@@ -41,9 +41,16 @@ int has_common_chars(const char* str1, const char* str2) {
 
 int main() {
     char input[10000];
-    std::cin.getline(input, sizeof(input));
+    if (!std::cin.getline(input, sizeof(input))) {
+        std::cerr << "Ошибка ввода\n";
+        return 1;
+    }
 
     char* result_rmv = static_cast<char*>(malloc(strlen(input) + 1));
+    if (result_rmv == nullptr) {
+        std::cerr << "Ошибка выделения памяти\n";
+        return 1;
+    }
     em::remove_latin(input, result_rmv);
     std::cout << result_rmv << std::endl;
     free(result_rmv);
